some changes

+++ app/Http/Controllers/Controller.php
@@ -22,19 +22,21 @@ class Controller extends BaseController
 
     public function afterPayment($uuid, $method)
     {
-        $order = Order::with('products')->where('uuid', $uuid)->whereNull('status')->firstOrFail();
-        $order->payment_method = $method;
-        $order->status = Order::PROCESSING;
-        $order->save();
+        $order = Order::with('products')->where('uuid', $uuid)->firstOrFail();
+        if ($order->status === null) {
+            $order->payment_method = $method;
+            $order->status = Order::PROCESSING;
+            $order->save();
 
-        Coupon::where('id', $order->coupon_id)->update(['used' => true]);
+            Coupon::where('id', $order->coupon_id)->update(['used' => true]);
 
-        Bag::where(auth()->check() ? ['user_id' => auth()->id()] : ['cookie_id' => $request->cookie('cookie_id')])->delete();
+            Bag::where(auth()->check() ? ['user_id' => auth()->id()] : ['cookie_id' => $request->cookie('cookie_id')])->delete();
 
-        foreach ($order->products as $item) {
-            $size = ProductSize::where('product_id', $item->product_id)->where('size_id', $item->size_id)->first();
-            $size->qty = $size->qty - $item->qty;
-            $size->save();
+            foreach ($order->products as $item) {
+                $size = ProductSize::where('product_id', $item->product_id)->where('size_id', $item->size_id)->first();
+                $size->qty = $size->qty - $item->qty;
+                $size->save();
+            }
         }
     }
